+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Pending
 
+* [AUI] Create Seperate File for Experiment EndScreen Tests [#15814](https://github.com/Applifier/unity-ads-webview/pull/15814)
+* [AUI] Rename AnimatedDownloadButtonEndScreen and its related files to ExperimentEndScreen [#15813](https://github.com/Applifier/unity-ads-webview/pull/15813)
 * [Load V5] Skip invalidation for no fill in RELOAD request [#15709](https://github.com/Applifier/unity-ads-webview/pull/15709)
 * Merge 3.4.6 to Master [#15784](https://github.com/Applifier/unity-ads-webview/pull/15784)